Reworked INI entirely and converted to an interface
A lot of changes to the Ini parser to make it slower on initializing but much faster on reading and cleaning it up overall. Most likely systems around it are broke and will need loose ends tied up.

Added a Converter Interface also Conv.h to handle all conversions in a standardized way. Still have to move some of the encoding string conversions over to it but this will need some testing.