Added AtlasUID FormUID DuetUID
Decided a U32 would suffice for atlas UID however I feel it's a bit tight and in the future might need to be increased to a U64. U64 requires a lot of focus and additional oversite to manage handing it back and forth to the UI and parsing in JSON without it getting abbreviated to a U32 so these wrappers will greatly aid in refactoring in the future if these int sizes have to be changed.