Changed UID for Atlas/Form to S32
Swapped Atlas Form to int 32 signed. Even though we're losing 50% of the storage cost  by not using negative values perhaps the negative range will come in handy in the future. Using U32 required to many conversions to strings when talking to other systems and most likely isn't worth the overhead.